+++ django_valkey/async_cache/client/__init__.py
@@ -1,4 +1,4 @@
-from django_valkey.async_cache.client.default import AsyncDefaultClient
+from django_valkey.async_cache.client.default import AsyncDefaultClient, AsyncHashClient
 from django_valkey.async_cache.client.herd import AsyncHerdClient
 from django_valkey.async_cache.client.sentinel import AsyncSentinelClient