+++ django_valkey/client/__init__.py
@@ -1,4 +1,4 @@
-from django_valkey.client.default import DefaultClient
+from django_valkey.client.default import DefaultClient, HashClient
 from django_valkey.client.herd import HerdClient
 from django_valkey.client.sentinel import SentinelClient
 from django_valkey.client.sharded import ShardClient